Can click on block and have it affect sprite 1
Run a script in the execution engine, and connect to it from Snap.  If
you click on a block (like 'turn 15 degrees') the execution does it to
the first sprite (and returns a result if there was one).  [The result
doesn't show up in the UI yet, but is logged to the javascript
console].  While there is a fair bit of spew to the console/terminal, I
thought I'd check this in anyways.